+++ lib/helpers_plethora.dart
@@ -3,3 +3,4 @@ library helpers_plethora;
 export 'src/extensions/datetime_x.dart';
 export 'src/extensions/datetimerange_x.dart';
 export 'src/extensions/list_datetimerange_x.dart';
+export 'src/extensions/string_x.dart';