+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0-nullsafety.1] - 2021-1-7.
+
+* Added `lastMonthDay` as extension getter method for `DateTime`.
+* Added extension methods for `String` class.
+
 ## [0.1.0-nullsafety.1] - 2021-1-7.
 
 * First pre-release.